refactor: update-file-structure-for-dashboard
